requirements.txt
asgiref==3.8.1
Django==5.2.3
django-debug-toolbar==5.2.0
django-redis==5.4.0
djangorestframework==3.16.0
Faker==37.4.0
pillow==11.2.1
pyotp==2.9.0
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn

